Add PvP Season APIs for Classic WoW (#50)

+++ wowp/characterEquipment.go
@@ -142,7 +142,7 @@ type CharacterEquipmentSummary struct {
 					R int     `json:"r"`
 					G int     `json:"g"`
 					B int     `json:"b"`
-					A float64 `json:"a"`
+					A float32 `json:"a"`
 				} `json:"color"`
 			} `json:"display"`
 		} `json:"armor,omitempty"`
